Fix for bug 1110526 (AMD64 fixes) by SF user galakhov.
git-svn-id: file:///.data/svn_convert/svn@5668 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/listview.h
@@ -87,7 +87,7 @@ public:
     static bool canDecode(QMimeSource*);
     static Contact *decode(QMimeSource*);
 protected:
-    unsigned m_id;
+    unsigned long m_id;
 };
 
 #endif
